pyproject.toml
[project]
name = "oz-externship-05"
version = "0.1.0"
description = "오즈 코딩 스쿨 초격차 익스턴십 5기 백엔드"
authors = [
    {name = "Meoyong",email = "dylim426@nextrunners.co.kr"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "sentry-sdk[django] (>=2.29.1,<3.0.0)",
    "django-redis (>=5.4.0,<6.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-filter (>=25.1,<26.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "twilio (>=9.8.4,<10.0.0)",
    "tblib (>=3.2.0,<4.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "boto3 (>=1.40.65,<2.0.0)",
    "channels (>=4.3.1,<5.0.0)",
    "channels-redis (>=4.3.0,<5.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "mypy-boto3-s3 (>=1.42.10,<2.0.0)",
    "google-genai (>=1.56.0,<2.0.0)",
]

[tool.poetry.group.dev.dependencies]
djangorestframework-stubs = "^3.16.0"
drf-spectacular = "^0.28.0"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.16.0"
coverage = "^7.8.2"
django-debug-toolbar = "^5.2.0"
django-stubs = "^5.2.0"
django-filter-stubs = "^0.1.3"
freezegun = "^1.5.5"
pandas-stubs = "^2.3.2.250926"
scipy-stubs = "^1.16.2.4"
boto3-stubs = {extras = ["s3"], version = "^1.40.64"}
botocore-stubs = "^1.40.65"
moto = "^5.1.16"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile="black"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = 3.12
strict = true
exclude = ["migrations", "manage.py"]

[tool.django-stubs]
django_settings_module = "config.settings.local"

[tool.coverage.run]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/test_*.py",
    "*/settings/*",
    "*/manage.py",
    "*/urls.py",
    "*/wsgi.py",
    "*/asgi.py",
    "*testcase.py",
]

[tool.coverage.report]
fail_under = 80
show_missing = true
